==> notebooks/01_data_exploration_auto.ipynb <==
# EDA
Automatic EDA using ydata-profiling

# User params

file_name_review = "Subscription_Boxes_review.json"
file_name_metadata = "Subscription_Boxes_metadata.json"
max_samples = 10_000

# Imports

import os
import pandas as pd
from ydata_profiling import ProfileReport

base_folder = os.path.join('..', 'data', 'raw')
output_folder = os.path.join('..', 'data', 'output')

# create paths
review_file_path = os.path.join(base_folder, file_name_review)
metadata_file_path = os.path.join(base_folder, file_name_metadata)

review_profiling_path = os.path.join(output_folder, "review_profiling.html")
metadata_profiling_path = os.path.join(output_folder, "metadata_profiling.html")

# read json file into dataframe
df_metadata = pd.read_json(metadata_file_path)
df_review = pd.read_json(review_file_path)

# display
display(df_metadata.head(1))
display(df_review.head(1))


# Report

# review report
profile_review = ProfileReport(df_review.head(max_samples), title="Profiling Report for reviews")
profile_review.to_file(review_profiling_path)

# metadata report
profile_metadata = ProfileReport(df_metadata.head(max_samples), title="Profiling Report for metadata")
profile_metadata.to_file(metadata_profiling_path)
