==> notebooks/02_data_merge_and_splitting.ipynb <==
# Data Splitting

### Possible Strategies

1. **Random Split**:
    - **Description**: Randomly splits the dataset into training, validation, and test sets.
    - **Use Case**: General-purpose model training and evaluation.
    - **Pros**: Simple to implement, ensures that each split is representative of the overall dataset.
    - **Cons**: May not capture temporal or user/item-specific trends.

2. **Temporal Split**:
    - **Description**: Splits the dataset based on timestamps, using older reviews for training and more recent reviews for validation and testing.
    - **Use Case**: Evaluating models on their ability to generalize to future data.
    - **Pros**: Mimics real-world scenarios where future data is unknown, captures temporal trends.
    - **Cons**: May result in imbalanced splits if the dataset has a seasonal trend or uneven distribution of reviews over time.
    - **Splitting Strategy**: Given a chronological user interaction sequence of length N:
        - **Training part**: Item interactions with timestamp range (-∞, t_1).
        - **Validation part**: Item interactions with timestamp range [t_1, t_2).
        - **Testing part**: Item interactions with timestamp range [t_2, +∞).

3. **User-Based Split**:
    - **Description**: Ensures that reviews from the same user are only present in one of the training, validation, or test sets.
    - **Use Case**: Personalization and recommendation systems where user history is critical.
    - **Pros**: Prevents data leakage from user behavior patterns.
    - **Cons**: May lead to splits that are less representative of the overall data distribution.

4. **Item-Based Split**:
    - **Description**: Ensures that reviews for the same product are only present in one of the training, validation, or test sets.
    - **Use Case**: Evaluating models on new products that were not seen during training.
    - **Pros**: Tests the model's ability to generalize to unseen items.
    - **Cons**: Similar to user-based split, it may result in less representative splits.

5. **Stratified Split**:
    - **Description**: Ensures that the splits maintain the same distribution of a certain feature, such as rating or product category.
    - **Use Case**: Ensuring that the model performs well across different subsets of the data.
    - **Pros**: Maintains the distribution of important features, leading to more balanced splits.
    - **Cons**: More complex to implement, may still miss temporal or user/item-specific trends.

6. **Cross-Validation Split**:
    - **Description**: Uses k-fold cross-validation to create multiple training and validation splits.
    - **Use Case**: Robust model evaluation and hyperparameter tuning.
    - **Pros**: Provides a more comprehensive evaluation by using multiple data splits.
    - **Cons**: Computationally intensive, not suitable for large datasets if computational resources are limited.

7. **Leave Last Out Split**:
    - **Description**: A data-splitting strategy to pick up the latest two item interactions for evaluation. This strategy is widely used in many recommendation papers.
    - **Splitting Strategy**: Given a chronological user interaction sequence of length N:
        - **Training part**: The first N-2 items.
        - **Validation part**: The (N-1)-th item.
        - **Testing part**: The N-th item.

# User parameters

file_name_review = "Subscription_Boxes_review.json"
file_name_metadata = "Subscription_Boxes_metadata.json"

SPLITTING_STRATEGY = 'random'  # 'random', 'temporal', 'leave-last-out'

# Processing

import os, sys
sys.path.append(os.path.abspath('../src'))
from src.config import BASE_PATH_DATA, PATH_PROJECT
os.chdir(PATH_PROJECT)

import pandas as pd
from src.utils.io import load_dataframe
from src.config import BASE_PATH_DATA
from sklearn.model_selection import train_test_split

# read json file into dataframe
df_metadata = load_dataframe(os.path.join(BASE_PATH_DATA, 'raw', file_name_review))
df_review = load_dataframe(os.path.join(BASE_PATH_DATA, 'raw', file_name_metadata))

# Merge the datasets on 'parent_asin' with suffixes for duplicate columns
merged_df = pd.merge(df_review, df_metadata, 
                     on="parent_asin", how="inner", suffixes=("_review", "_metadata"))

# drop unnecessary columns (note user ID dropped later)
columns_to_drop = [
    "parent_asin", 
    "asin", 
    "images_review",  # too complex
    "images_metadata",  # too complex
    "videos",  # too complex
    "details",  # mostly empty
    "categories",  # mostly empty
    "bought_together",  # mostly empty
    "price",  # mostly empty
    "description",  # mostly empty
    ]
merged_df = merged_df.drop(columns=columns_to_drop)  # those features will be OOD at inference for new data

# Save the combined dataset
file_name_merged = os.path.join(BASE_PATH_DATA, 'processed', 'merged_dataset.parquet')
merged_df.to_parquet(file_name_merged, index=False)

# split the data
match SPLITTING_STRATEGY:
    case 'random':
        # Split the data into train and test sets
        train_df, test_df = train_test_split(merged_df, test_size=0.2, random_state=42)
    case 'temporal':
        # Perform temporal split
        # Use precomputed datasets
        raise ValueError("not implemented")
    case 'leave-last-out':
        # Perform leave-last-out split
        # Use precomputed datasets
        raise ValueError("not implemented")
    case _:
        raise ValueError(f"Unknown splitting strategy: {SPLITTING_STRATEGY}")
    
# save the train and test sets
train_df.to_parquet(os.path.join(BASE_PATH_DATA, 'processed', 'train.parquet'), index=False)
test_df.to_parquet(os.path.join(BASE_PATH_DATA, 'processed', 'test.parquet'), index=False)
