==> notebooks/05_model_serving.ipynb <==
# User params

n_data = 2  # numebr of samples

import os, sys
sys.path.append(os.path.abspath('../src'))
from src.config import BASE_PATH_DATA, PATH_PROJECT
os.chdir(PATH_PROJECT)

# Utility functions

# Prepare some data

from src.utils.io.io import load_dataframe
from src.config import BASE_PATH_DATA
from src.api.utils import create_batch_detection_request

# load test data
df_test = load_dataframe(os.path.join(BASE_PATH_DATA, 'processed', 'test.parquet'))
df_test = df_test.sample(n_data)  # take some samples
display(df_test)

# Print column names and their data types
for col in df_test.columns:
    print(f"{col}: {df_test[col].dtype}")

# format the data for a request
json_data = df_test.to_dict(orient="records")
payload = create_batch_detection_request(json_data)

print(payload.requests[0])
print(payload)


# Launch the service

**In local**

%%sh
# check process is at project root (
# -> ok because changed dir at the beginning of the notebook
pwd

# Kill any existing Uvicorn processes
kill $(ps aux | grep 'uvicorn' | grep -v grep | awk '{print $2}')

# Start a new Uvicorn server process with nohup to detach it and run it in the background
# poetry run uvicorn src.app:app --host 0.0.0.0 --port 8000 --reload: Run the Uvicorn server using Poetry
# --host 0.0.0.0: Bind the server to all network interfaces
# --port 8000: Use port 8000 for the server
# --reload: Enable auto-reloading for code changes (useful during development)
# > logs/uvicorn.log 2>&1: Redirect both stdout and stderr to the log file logs/uvicorn.log
# &: Run the command in the background
nohup poetry run uvicorn src.app:app --host 0.0.0.0 --port 8000 --reload > logs/uvicorn.log 2>&1 &

# Wait for 5 seconds to ensure the server has started
sleep 5

# Check if the server is running by sending a request
# If curl fails (i.e., the server is not running or the endpoint is incorrect), it outputs "Server is not running".
curl -f http://localhost:8000/ || echo "Server is not running"



**Using docker**

%%sh

# Kill any existing Uvicorn processes
kill $(ps aux | grep 'uvicorn' | grep -v grep | awk '{print $2}')

# build the container
make build_docker

# run it
make run_docker

# Wait for 5 seconds to ensure the server has started
sleep 5

# Check if the server is running by sending a request
# If curl fails (i.e., the server is not running or the endpoint is incorrect), it outputs "Server is not running".
curl -f http://localhost:8000/ || echo "Server is not running"

# Display the docs in the notebook (rather than in the browser) - for fun

# from IPython.display import IFrame

# # URL for the Swagger UI
# swagger_ui_url = "http://localhost:8000/docs"

# # Display Swagger UI in the notebook
# IFrame(src=swagger_ui_url, width='90%', height='600px')

# Query the service

**Using python**

import requests

# root query
headers = {
    "accept": "application/json",
    "Content-Type": "application/json"
}
response = requests.get("http://localhost:8000/", "", headers=headers)
print('Status Code:', response.status_code)
print('Response Text:', response.text)

import requests
# outlier query
url = 'http://localhost:8000/anomaly/detect_outliers'
headers = {
    "accept": "application/json",
    "Content-Type": "application/json"
}
# convert input to json (fastapi will convert it automatically to proper format using pydantic. no need to serialize)
response = requests.post(url, json=payload.dict(), headers=headers)

# Check the status code and print appropriate messages
print('Status Code:', response.status_code)
if response.status_code == 200:
    try:
        print(response.json())
    except requests.exceptions.JSONDecodeError:
        print("Response content is not valid JSON")
else:
    print(f"Request failed with status code {response.status_code}")
    print(response.text)
